Revert "Add Example console search"
This reverts commit 3557e75890910362fc2a2e4dcce447ad9cc7d794.
