+++ example/src/main/java/io/radar/example/console/ConsoleView.kt
@@ -12,37 +12,49 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Clear
 import androidx.compose.material.icons.filled.ContentCopy
 import androidx.compose.material.icons.filled.DeleteOutline
 import androidx.compose.material.icons.filled.Inbox
-import androidx.compose.material.icons.filled.Search
 import androidx.compose.material.icons.filled.Share
 import androidx.compose.material3.FilterChip
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.unit.dp
+import io.radar.example.store.ConsoleEntry
+import io.radar.example.store.ConsoleKind
 import io.radar.example.store.LocalLogStore
 import android.content.ClipData
 import android.content.ClipboardManager
+import java.text.SimpleDateFormat
+import java.util.Date
+import java.util.Locale
 
+private enum class ConsoleFilter(val label: String, val kinds: Set<ConsoleKind>?) {
+    ALL("All", null),
+    ACTIONS("Actions", setOf(ConsoleKind.ACTION, ConsoleKind.RESULT)),
+    EVENTS("Events", setOf(ConsoleKind.EVENT)),
+    LOCATIONS("Locations", setOf(ConsoleKind.LOCATION)),
+    LOGS("Logs", setOf(ConsoleKind.LOG)),
+    ERRORS("Errors", setOf(ConsoleKind.ERROR)),
+    ;
+
+    fun matches(kind: ConsoleKind): Boolean = kinds?.contains(kind) ?: true
+}
+
+private val exportTimeFormat = SimpleDateFormat("HH:mm:ss.SSS", Locale.US)
 
 /** The Debug tab: a unified, filterable console timeline over [LogStore.entries]. */
 @Composable
@@ -50,10 +62,9 @@ fun ConsoleView(modifier: Modifier = Modifier) {
     val logStore = LocalLogStore.current
     val context = LocalContext.current
     var filter by remember { mutableStateOf(ConsoleFilter.ALL) }
-    var searchText by rememberSaveable { mutableStateOf("") }
 
     val entries = logStore.entries
-    val filtered = filterConsoleEntries(entries, filter, searchText)
+    val filtered = entries.filter { filter.matches(it.kind) }
 
     Column(modifier = modifier.fillMaxSize()) {
         // Header
@@ -69,63 +80,30 @@ fun ConsoleView(modifier: Modifier = Modifier) {
                 fontWeight = FontWeight.Bold,
             )
             Text(
-                text = "  ${filtered.size}",
+                text = "  ${entries.size}",
                 style = MaterialTheme.typography.bodyMedium,
                 color = MaterialTheme.colorScheme.onSurfaceVariant,
                 modifier = Modifier.weight(1f),
             )
-            IconButton(
-                enabled = filtered.isNotEmpty(),
-                onClick = {
-                    val manager = context.getSystemService(ClipboardManager::class.java)
-                    manager?.setPrimaryClip(ClipData.newPlainText("Radar console", exportConsoleText(filtered)))
-                },
-            ) {
+            IconButton(onClick = {
+                val manager = context.getSystemService(ClipboardManager::class.java)
+                manager?.setPrimaryClip(ClipData.newPlainText("Radar console", exportText(entries)))
+            }) {
                 Icon(Icons.Filled.ContentCopy, contentDescription = "Copy")
             }
-            IconButton(
-                enabled = filtered.isNotEmpty(),
-                onClick = {
-                    val send = Intent(Intent.ACTION_SEND).apply {
-                        type = "text/plain"
-                        putExtra(Intent.EXTRA_TEXT, exportConsoleText(filtered))
-                    }
-                    context.startActivity(Intent.createChooser(send, "Share console"))
-                },
-            ) {
+            IconButton(onClick = {
+                val send = Intent(Intent.ACTION_SEND).apply {
+                    type = "text/plain"
+                    putExtra(Intent.EXTRA_TEXT, exportText(entries))
+                }
+                context.startActivity(Intent.createChooser(send, "Share console"))
+            }) {
                 Icon(Icons.Filled.Share, contentDescription = "Share")
             }
-            IconButton(
-                enabled = entries.isNotEmpty(),
-                onClick = { logStore.clear() },
-            ) {
+            IconButton(onClick = { logStore.clear() }) {
                 Icon(Icons.Filled.DeleteOutline, contentDescription = "Clear")
             }
         }
-
-        OutlinedTextField(
-            value = searchText,
-            onValueChange = { searchText = it },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 12.dp, vertical = 4.dp),
-            label = { Text("Search logs") },
-            leadingIcon = {
-                Icon(Icons.Filled.Search, contentDescription = null)
-            },
-            trailingIcon = {
-                if (searchText.isNotEmpty()) {
-                    IconButton(onClick = { searchText = "" }) {
-                        Icon(Icons.Filled.Clear, contentDescription = "Clear search")
-                    }
-                }
-            },
-            singleLine = true,
-            keyboardOptions = KeyboardOptions(
-                capitalization = KeyboardCapitalization.None,
-                autoCorrectEnabled = false,
-            ),
-        )
 
         // Filter chips
         Row(
@@ -154,24 +132,11 @@ fun ConsoleView(modifier: Modifier = Modifier) {
                         contentDescription = null,
                         tint = MaterialTheme.colorScheme.onSurfaceVariant,
                     )
-                    if (entries.isEmpty()) {
-                        Text(
-                            text = "Nothing yet",
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                    } else {
-                        Text(
-                            text = "No matches",
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                        Text(
-                            text = "Change the filter or search term.",
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                    }
+                    Text(
+                        text = "Nothing yet",
+                        style = MaterialTheme.typography.bodyMedium,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant,
+                    )
                 }
             }
         } else {
@@ -184,3 +149,10 @@ fun ConsoleView(modifier: Modifier = Modifier) {
         }
     }
 }
+
+private fun exportText(entries: List<ConsoleEntry>): String =
+    entries.joinToString("\n") { entry ->
+        val time = exportTimeFormat.format(Date(entry.timestamp))
+        val head = "[$time] ${entry.kind.name} — ${entry.summary}"
+        if (entry.detail.isNullOrBlank()) head else "$head\n${entry.detail}"
+    }
